// src/app/models/course.ts
export class Course {

    id: number | any;

    name: string | undefined;

    summary: string | undefined;

    description: string | undefined;

    difficultyType: number | undefined;

    unitPrice: number | undefined;

    imageUrl: string | undefined;

    demoUrl: string | undefined;

    url: string | undefined;

    categoryId: number | undefined;

    mentorId: number | undefined;

    sequence : string | undefined;

    isActive : boolean | undefined;

}